# precompile: skip small functions