These are likely to never be called directly, and would be quick to compile if they happened to be. In v0.6, I used to be manually excluding many of these from the sysimg; this is just a more general.

before
2565541888  maximum resident set size # 2.5 GB

after
1980424192  maximum resident set size # 2.0 GB

(build time and .ji size are also about proportionally (~25%) improved)