@KristofferC 2018-02-24
Running with Pkg3 precompilation on, I get the following with GNU Time:

Before PR:
```
   Command being timed: "make"
        Elapsed (wall clock) time (h:mm:ss or m:ss): 6:43.50
        Maximum resident set size (kbytes): 3259596
```

After PR:
```
 Command being timed: "make"
        Elapsed (wall clock) time (h:mm:ss or m:ss): 5:59.28
        Maximum resident set size (kbytes): 2551840
```

There seem to be some slight regressions in first call timing 0.5s -> 0.8s for first call to `status` and 1.5s -> 1.8s for first call to `add` so not that bad.