@KristofferC 2018-02-24
FWIW, current master is at `Maximum resident set size (kbytes): 2843524`. I haven't tried this PR without Pkg3 precompile yet.