@staticfloat 2018-02-25
Verification runs on buildbots:

ubuntu32: https://build.julialang.org/#/builders/68/builds/817 (this one doesn't really matter, it's just as verification that failures on x86 aren't due to something weird about the old CentOS images we use for the `linux32` builder)
linux32: https://build.julialang.org/#/builders/47/builds/811 
win32: https://build.julialang.org/#/builders/44/builds/819