@staticfloat 2018-02-25
Confirmed that this PR passes tests on all the buildbots, tragically the win32 buildbot fails on `make docs` after passing all tests.