@staticfloat 2018-02-25
For some reason it's saved into `docbuild.log` instead of just written out, but it's TLS issues while trying to clone `METADATA.jl`.