@staticfloat 2018-02-25
I applied the hotfix to the windows buildbots, I'm re-running the win32 build of this branch here: https://build.julialang.org/#/builders/44/builds/821