@StefanKarpinski 2018-04-13
Based on the analysis @KristofferC did in #26767, the premise of this PR seems like it may be false.